chore: restructure class folder structure
