+++ app/Models/Core/Challenge.php
